cleanup module structure
