--- iroh-willow/src/session/aoi_finder.rs
@@ -1,16 +1,0 @@
-// use std::{cell::RefCell, collections::VecDeque, rc::Rc, task::Waker};
-//
-// use crate::actor::AreaOfInterestPair;
-
-// pub struct AoiFinder {
-//     inner: Rc<RefCell<Inner>>
-// }
-//
-// impl AoiFinder {
-//     fn push
-// }
-//
-// struct Inner {
-//     queue: VecDeque<AreaOfInterestPair>,
-//     wakers: VecDeque<Waker>
-// }
